ovl: make directory ino/dev numbers stable
Overlayfs reported pseudo ino/dev in stat(2) for dirctories.  This caused
problems when in tight memory situations dir dentries were evicted from the
cache and recreated with a different inode number:

  "We're receiving bug reports from users simply untarring archives onto
   overlayfs mounts (docker + overlayfs) where tar fails with "Directory
   renamed before its status could be extracted" errors.

   This is triggered in GNU tar when the inode number of an ancestor
   directory changes unexpectedly in the deferred application of ancestral
   metadata."

To fix this, use the lower ino + overlayfs dev number for lower and merged
directories and the upper ino/dev for non-merged upper directories.

This scheme allow the behavior to confirm to the following criteria:

 - ino/dev is stable across copy-up
 - ino/dev is stable for the lifetime of the mount (*)
 - ino/dev uniquely identifies objects on the system

It does not confirm to

 - ino/dev is stable after umount+mount
 - ino/dev is the same for all directories under the overlay fs (*)

The ones marked with (*) are those that this patch changes.  This last one
has the potential to cause regressions.  In particular it will make
unionmount-testsuite fail, so needs the following change to context.py:

<             if dev != self.upper_dir_fs():
<                 raise TestError(name + ": Directory not on union layer")
>             pass

Reported-by: Vito Caputo <vito.caputo@coreos.com>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>