PCI: rockchip: Add quirk to disable RC's ASPM L0s
Rockchip's RC outputs 100MHz reference clock but there are two methods for
PHY to generate it.

(1) One of them is to use system PLL to generate 100MHz clock and the PHY
will relock it and filter signal noise then output the reference clock.

(2) Another way is to share Soc's 24MHZ crystal oscillator with PHY and
force PHY's DLL to generate 100MHz internally.

When using case (2), the exit from L0s occasionally doesn't work due to the
broken design of RC receiver's logical circuit.  So even if we use
extended-synch, it still fails for PHY to relock the bits from FTS
sometimes.  This will hang the system.

Maybe we could argue that why not use case (1) to avoid it?  The reason is
that as we could see the reference clock is derived from system PLL and the
path from it to PHY isn't so clean which means there are some noise
introduced by power-domain and other buses can't be filterd out by PHY and
we could see noise from the frequency spectrum by oscilloscope.  This makes
the TX compatibility test a little difficult to pass the spec.  So case (1)
and case (2) are both used.  If using case (2), we should disable RC's L0s
support, and that is why we need this property to indicate this quirk.

Reported-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Signed-off-by: Shawn Lin <shawn.lin@rock-chips.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
CC: Brian Norris <briannorris@chromium.org>