vfs: mnt_want_write_file() should freeze protect underlying sb
If mnt_want_write_file() is called from filesystem code (which is the vast
majority of the cases), then we get freeze protection on the overlay but
not on the underlying filesystem.

This is wrong, we do need freeze protection on the unerlying fs and don't
need it on the overlay.

Having freeze protection on the overlay doesn't hurt, but currently there's
no way to freeze the overlay so this issue is moot.

There are a small number of exceptions where we want to preserve the old
behavior of mnt_want_write_file(): fchown, fsetxattr, fremovexattr.  These
get freeze protection on the underlying fs inside the relevant overlayfs
i_op.

This patch is a no-op for anything other then overlayfs.

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>