mm: Prevent __alloc_pages_nodemask() RCU CPU stall warnings
This commit adds a cond_resched_rcu_qs() call in __alloc_pages_nodemask()
(which is invoked indirectly from alloc_page_interleave()) in order to
prevent RCU CPU stall warnings.  More importantly, this change allows
RCU grace periods to complete despite "longer than average" loops in
__alloc_pages_nodemask().

Note that this is a case of "treating the symptoms".  In addition,
it is quite possible that the cond_resched_rcu_qs() needs to be
in some lower-level function that might have been inlined into
__alloc_pages_nodemask().  In short, this one is experimental in nature.

Reported-by: Boris Zhmurov <bb@kernelpanic.ru>
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Cc: Michal Hocko <mhocko@kernel.org>
Cc: Andrew Morton <akpm@linux-foundation.org>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: <linux-mm@kvack.org>