mm/thp/pagecache: only withdraw page table after a successful deposit
The current code wrongly called withdraw in the error path.  But we
haven't depoisted the page table yet in the only error path in that
function.  So for now remove that withdraw completely.  If we take that
"out:" branch, we should have vmf->prealloc_pte already pointing to the
allocated page table.

Fixes: 953c66c2b22a304d ("mm: THP page cache support for ppc64")
Link: http://lkml.kernel.org/r/20161212163428.6780-1-aneesh.kumar@linux.vnet.ibm.com
Signed-off-by: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>
Reported-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
Acked-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>