svcauth_gss: free context promptly on PROC_DESTROY
Since c5b29f885afe "sunrpc: use seconds since boot in expiry cache", the
expiry_time field has been in units of seconds since boot, but that
patch overlooked the server code that handles RPC_GSS_PROC_DESTROY
requests.  The result is that when we get a request from a client to
destroy a context, we set that context's expiry time to a time decades
in the future.

The context will still be cleaned up by cache_purge when the module is
unloaded or the container shut down, but a lot of contexts could pile up
before then.

The simple fix would be just to set expiry_time to seconds_since_boot(),
but modifying the entry outside the cache_lock is also looks dubious
(though I'm not completely sure what the race would be).  So let's
provide a helper that does this properly, taking the lock and removing
the entry immediately.

Signed-off-by: Neil Brown <neilb@suse.com>
Reported-by: Andy Adamson <andros@netapp.com>
Signed-off-by: Andy Adamson <andros@netapp.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>