parisc: Enhance detection if cr16 clocksources are synchronous
The cr16 clocks of the CPUs are usually not synchron.  Nevertheless, it
seems that each socket (with two cores) of a C8000 workstation is fed by
the same clock source, which makes the cr16 clocks of each socket
syncron. Let's try to detect such situations and mark the cr16
clocksource and the sched_clock stable on single-socket and single-core
machines.

Signed-off-by: Helge Deller <deller@gmx.de>