rcu: Make synchronize_rcu_expedited() safe for early boot
The synchronize_rcu_expedited() function does not check for early-boot
use, which can result in failures if it is invoked before the scheduler
has started.  Given that the rcupdate.rcu_expedited kernel parameter
causes all calls to synchronize_rcu() to be directed instead to
synchronize_rcu_expedited(), a usage restriction does not make sense.

This commit therefore adds a rcu_scheduler_active check to
synchronize_rcu_expedited(), so that it is a no-op before the scheduler
starts.  This behavior is correct because there is only a single CPU
running during that time.

Reported-by: Lv Zheng <lv.zheng@intel.com>
Reported-by: Borislav Petkov <bp@alien8.de>
Fixes: 174cc7187e6f ("ACPICA: Tables: Back port acpi_get_table_with_size() and early_acpi_os_unmap_memory() from Linux kernel")
Link: Link: https://lkml.kernel.org/r/4034dde8-ffc1-18e2-f40c-00cf37471793@intel.com
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>