rcu: Add early boot testing of synchronous grace-period primitives
The various synchronize_*() synchronous grace-period primitives are
supposed to be legal at early boot, but this is not tested, and as a
result synchronize_rcu_expedited() failed to implement this requirement.
In order to prevent future failures of this sort, this commit adds early
boot testing of these primitives when CONFIG_PROVE_RCU is enabled.

Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>