clk: meson-gxbb: Add MALI clocks and fix GP0 pll support
The Mali is clocked by two identical clock paths behind a glitch free mux
to safely change frequency while running.

The optimal clock frequency is provided by the GP0 PLL, but unlike Meson8 the
GXBB GP0 PLL needs some tweaking and need some SoC specific values to be set
in order to setup the PLL.

Since these values will change for GXL and GXM, add an init table that will
be changed for GXL or GXM.

For the MALI clocks, introduce them as Composite clocks to simplify the
management.
To handle composite clocks, a gxbb_composite structure is added to dynamically
create the composite clock at probe and fill accordingly the DT table.

Finally the glitch free mux is added and other muxes and dividers are also
grouped in initialization tables.