genirq: Reenable shared irq debugging in request_*_irq()
Commit 6d83f94db95c ("genirq: Disable the SHIRQ_DEBUG call in
request_threaded_irq for now") disabled the SHIRQ_DEBUG facility for the
irq request side. Unfortunately the 'for now' part and the _FIXME add on to
the CONFIG_SHIRQ_DEBUG guard got forgotten.

It's unfortunately not possible to keep the original idea of just invoking
the newly installed handler due to the reasons described in the commit
which disabled it, but having this debug facility available makes a lot of
sense.

The only way to regain that testing is to invoke the real interrupt
handling mechanism, which will also invoke an already registered handler,
but that's not an issue because _all_ interrupt handlers need to be
prepared for handling spurious interrupts.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: Peter Zijlstra <peterz@infradead.org>
Link: http://lkml.kernel.org/r/alpine.DEB.2.20.1702151542330.3562@nanos
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>