leds: core: use deferred probing if default trigger isn't available yet
When registering a LED device we have the option to set a default trigger.
Depending on load order of drivers this trigger may not be available yet.
(affected LED device in my case: a DT-configured GPIO LED)
So far if the default trigger can't be found this error is silently
ignored.

Let's change this to return EPROBE_DEFER if the default trigger can't be
found. This gives the system the chance to probe the LED device later
once the trigger is available.

In addition in led_trigger_set_default break out of the loop when trigger
has been found.

Signed-off-by: Heiner Kallweit <hkallweit1@gmail.com>
Signed-off-by: Jacek Anaszewski <jacek.anaszewski@gmail.com>